+++ src/executor_c.cpp
@@ -250,7 +250,7 @@ int executor_init_and_run_wait(executor_t exec) {
 
 int executor_stop(executor_t exec) {
     // std::cout << "executor_stop() - 1\n";
-    int res = exec->actual.stop();
+    int res = static_cast<int>(exec->actual.stop());
     // std::cout << "executor_stop() - 2\n";
     return res;
 }
